chore: remove logyaml
