+++ clientcmd/log.go
@@ -6,10 +6,10 @@ import (
 	"fmt"
 	"io"
 	"sort"
+	"strings"
 
 	"github.com/flanksource/clicky"
 	"github.com/flanksource/commons/logger"
-	"sigs.k8s.io/yaml"
 )
 
 func shouldLogJSON() bool {
@@ -29,19 +29,6 @@ func Log(w io.Writer, data map[string]any) error {
 	return err
 }
 
-func LogYAML(w io.Writer, data map[string]any) error {
-	if shouldLogJSON() {
-		return Log(w, data)
-	}
-
-	out, err := yaml.Marshal(data)
-	if err != nil {
-		return err
-	}
-	_, err = w.Write(out)
-	return err
-}
-
 func formatKeyValues(data map[string]any) []byte {
 	keys := make([]string, 0, len(data))
 	for key := range data {
@@ -57,4 +44,49 @@ func formatKeyValues(data map[string]any) []byte {
 		fmt.Fprintf(&out, "%s=%v", key, data[key])
 	}
 	return out.Bytes()
+}
+
+func printClicky(w io.Writer, data any, defaultFormat string) error {
+	opts := clicky.Flags.FormatOptions
+	if err := opts.ParseFormatSpec(); err != nil {
+		return err
+	}
+
+	if len(opts.Sinks) == 0 {
+		return writeClickyOutput(w, data, clicky.FormatOptions{Format: defaultFormat}, opts)
+	}
+
+	for _, sink := range opts.Sinks {
+		sinkOpts := opts
+		sinkOpts.Sinks = nil
+		sinkOpts.Format = sink.Format
+		sinkOpts.JSON, sinkOpts.YAML, sinkOpts.CSV = false, false, false
+		sinkOpts.HTML, sinkOpts.Markdown, sinkOpts.Pretty = false, false, false
+		sinkOpts.PDF, sinkOpts.Slack = false, false
+		if sink.File == "" {
+			if err := writeClickyOutput(w, data, sinkOpts); err != nil {
+				return err
+			}
+			continue
+		}
+		sinkOpts.Output = sink.File
+		if err := clicky.Formatter.FormatToFile(sinkOpts, data); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func writeClickyOutput(w io.Writer, data any, opts ...clicky.FormatOptions) error {
+	out, err := clicky.Format(data, opts...)
+	if err != nil {
+		return err
+	}
+	if _, err := fmt.Fprint(w, out); err != nil {
+		return err
+	}
+	if !strings.HasSuffix(out, "\n") {
+		_, err = fmt.Fprintln(w)
+	}
+	return err
 }
